requirements.txt
requests>=2.20.1
pyOpenSSL>=18.0.0
